django/django_projects/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
distlib==0.3.8
Django==5.0.4
djangorestframework==3.15.1
filelock==3.13.3
idna==3.7
pillow==10.3.0
platformdirs==4.2.0
requests==2.31.0
sqlparse==0.4.4
urllib3==2.2.1
virtualenv==20.25.1
drf-yasg==1.21.7
setuptools==70.0.0
